add_parser has too much code copy-pasted
